feat: Add functionality to remember the last viewed page
The code changes include adding functionality to remember the last viewed page in the `Default.vue` layout. When the user clicks on the explore button, the `viewDetails` function is called, which sets the value of `localStorage` to "discover" and navigates the user to the "/discover" route. In the `discover.vue` layout, the `close` function removes the "page" item from `localStorage`.

This commit message follows the established convention of using a verb in the imperative form ("Add functionality") and provides a clear and concise description of the changes made.