feat: Add onUnmounted hook to save last viewed page in Default.vue
The code changes in Default.vue include adding the onUnmounted hook to save the last viewed page in the localStorage. When the user navigates away from the Default.vue layout, the onUnmounted hook is triggered and sets the value of "page" in localStorage to "home". This ensures that when the user returns to the Default.vue layout, the last viewed page is remembered.

This commit message follows the established convention of using a verb in the imperative form ("Add onUnmounted hook") and provides a clear and concise description of the changes made.