feat: Save last viewed page in localStorage using onUnmounted hook
The code changes in Default.vue include adding the onUnmounted hook to save the last viewed page in the localStorage. When the user navigates away from the Default.vue layout, the onUnmounted hook is triggered and sets the value of "page" in localStorage to "home". This ensures that when the user returns to the Default.vue layout, the last viewed page is remembered.