refactor: Update animation values in discover.vue
Update the animation values in discover.vue to improve the animation effect when closing the discover layout. The x value is changed to "-800px" and the y value is changed to "-800px".

This commit message follows the established convention of using a verb in the imperative form ("Update animation values") and provides a clear and concise description of the changes made.