refactor: Update tooltip text and location in Default.vue and discover.vue
The code changes in Default.vue and discover.vue update the tooltip text and location. In Default.vue, the tooltip text is changed from "Profile" to "View Details" and the location is changed from "start" to "start,end". In discover.vue, the tooltip text remains the same but the location is changed from "start" and "end" to "start" and "end". These changes improve the clarity and usability of the tooltips in both layouts.