refactor: Update social media links for Ferasat
The code changes in ferasatSocialLinks.ts update the GitHub and Twitter links for Ferasat. The GitHub link is changed to "https://github.com/mferasatali" and the Twitter link is changed to "https://twitter.com/mferasat_ali". This commit improves the accuracy of the social media links for Ferasat.

This commit message follows the established convention of using a verb in the imperative form ("Update social media links") and provides a clear and concise description of the changes made.